feat: extract orchestration to VoiceCallService
