+++ VoiceCallAssistant/Interfaces/IVoiceCallService.cs
@@ -1,0 +1,8 @@
+using System.Net.WebSockets;
+
+namespace VoiceCallAssistant.Interfaces;
+
+public interface IVoiceCallService
+{
+    Task OrchestrateAsync(WebSocket websocket, string userPrompt, CancellationTokenSource cancellationTokenSource);
+}
